requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.3
django-admin-rangefilter==0.6.3
django-jazzmin==2.4.1
django-money==1.2
polib==1.1.0
py-moneyed==0.8.0
python-money==0.5
pytz==2020.4
sqlparse==0.4.1